Module apis::channel

Modules

backend

Structs

Channel
Main channel struct.

Def
Channel definition.

RecvError
Sender disconnected, no further messages will ever be received.

SendError
Receiver disconnected, message will never be deliverable.

Enums

CreateError
Error creating concrete Channel instance from a given channel def.

DefineError
Error defining Def.

Kind
Channel kind defines the connection topology of a channel.

TryRecvError

Traits

Endpoint
Interface for a channel endpoint.

Id
Unique identifier with a total mapping to channel infos.

Sourcepoint
Interface for a channel sourcepoint.

Functions

report_sizes