Struct apis::channel::Def

pub struct Def<CTX: Context> { /* fields omitted */ }

Channel definition.

Methods

impl<CTX: Context> Def<CTX>

pub fn define(
id: CTX::CID,
kind: Kind,
producers: Vec<CTX::PID>,
consumers: Vec<CTX::PID>
) -> Result<Self, Vec<DefineError>>

The only method for creating valid channel def struct; validates specification of sourcepoints and endpoints for well-formedness (at least one process at each end, no duplicates or self-loops) and compatibility with channel kind (restricted to single process sourcepoint or endpoint where appropriate).

Errors

Zero producers or consumers:

let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Sink,
  vec![],
  vec![ProcessId::B]);
assert_eq!(result, Err (vec![channel::DefineError::ZeroProducers]));
let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Sink,
  vec![ProcessId::A],
  vec![]);
assert_eq!(result, Err (vec![channel::DefineError::ZeroConsumers]));

Producer equals consumer:

let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Sink,
  vec![ProcessId::A, ProcessId::B],
  vec![ProcessId::A]);
assert_eq!(result, Err (vec![channel::DefineError::ProducerEqConsumer]));

Duplicate producer:

let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Sink,
  vec![ProcessId::A, ProcessId::A],
  vec![ProcessId::B]);
assert_eq!(result, Err (vec![channel::DefineError::DuplicateProducer]));

Duplicate consumer:

let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Source,
  vec![ProcessId::A],
  vec![ProcessId::B, ProcessId::B]);
assert_eq!(result, Err (vec![channel::DefineError::DuplicateConsumer]));

Kind does not support multiple producers and/or consumers:

let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Source,
  vec![ProcessId::A, ProcessId::B],
  vec![ProcessId::C, ProcessId::D]);
assert_eq!(result, Err (vec![channel::DefineError::MultipleProducers]));
let result = channel::Def::<Mycontext>::define (
  ChannelId::X,
  channel::Kind::Simplex,
  vec![ProcessId::A],
  vec![ProcessId::B, ProcessId::C]);
assert_eq!(result, Err (vec![channel::DefineError::MultipleConsumers]));

pub fn id(&self) -> &CTX::CID

pub fn kind(&self) -> &Kind

pub fn producers(&self) -> &Vec<CTX::PID>

pub fn consumers(&self) -> &Vec<CTX::PID>

pub fn to_channel<M>(self) -> Channel<CTX> where
CTX: 'static,
M: Message<CTX> + 'static,