Enum apis::process::Kind

pub enum Kind {
    Asynchronous {
        messages_per_update: u32,
    },
    Isochronous {
        tick_ms: u32,
        ticks_per_update: u32,
    },
    Mesochronous {
        tick_ms: u32,
        ticks_per_update: u32,
    },
    Anisochronous,
}

Specifies the loop behavior of a process.

Asynchronous is a loop that blocks waiting on exactly one channel endpoint.

Isochronous is a fixed-timestep loop in which endpoints are polled once per 'tick' and will attempt to "catch up" if it falls behind.

Mesochronous is a rate-limited loop that polls processes and loops immediately if enough time has passed and otherwise sleeps for the remaining duration until the next 'tick'.

Anisochronous is an un-timed polling loop which always loops immediately and always processes one update per tick.

Variants
Asynchronous

Block waiting on one or more endpoints.

Asynchronous processes can only hold multiple endpoints of compatible kinds of channels. Currently this is either any number of sink endpoints, or else any number and combination of simplex or source endpoints. This is validated internally when defining an Def struct with the provided kind and endpoints.

Fields of Asynchronous

messages_per_update: u32

Isochronous

A fixed-time step polling loop that will try to "catch up" if it falls behind.

Fields of Isochronous

tick_ms: u32

ticks_per_update: u32

Mesochronous

A rate-limited polling loop.

Fields of Mesochronous

tick_ms: u32

ticks_per_update: u32

Anisochronous

Poll to exhaustion and update immediately.

This is useful for blocking update functions as in a readline loop or a rendering loop.

Note that unlike other polling process kinds (Isochronous and Mesochronous), ticks and updates are always one-to-one since an external blocking mechanism is expected to be used in the update function.

Methods

impl Kind

pub fn asynchronous_default() -> Self

pub fn isochronous_default() -> Self

pub fn mesochronous_default() -> Self

pub fn anisochronous_default() -> Self

pub fn new_asynchronous(
messages_per_update: u32
) -> Result<Self, Vec<KindError>>

pub fn new_isochronous(
tick_ms: u32,
ticks_per_update: u32
) -> Result<Self, Vec<KindError>>

pub fn new_mesochronous(
tick_ms: u32,
ticks_per_update: u32
) -> Result<Self, Vec<KindError>>

pub fn new_anisochronous() -> Self