Trait apis::process::Id

pub trait Id<CTX> where
    Self: EnumUnitary,
    CTX: Context<PID = Self>, {
    fn def(&self) -> Def<CTX>;

    fn spawn(inner: Inner<CTX>) -> JoinHandle<Option<()>>;

    fn gproc(inner: Inner<CTX>) -> CTX::GPROC;
}

Unique identifier with a total mapping to process defs.

Required Methods

fn def(&self) -> Def<CTX>

fn spawn(inner: Inner<CTX>) -> JoinHandle<Option<()>>

Must initialize the concrete process type start running the initial closure.

fn gproc(inner: Inner<CTX>) -> CTX::GPROC

Initialie the concrete proces type and return in a CTX::GPROC.