Trait apis::session::Context

pub trait Context where
    Self: Clone + PartialEq + Sized + Debug, {
    type MID: Id;
    type CID: Id<Self>;
    type PID: Id<Self>;
    type GMSG: Global<Self>;
    type GPROC: Global<Self>;
    type GPRES: Global<Self>;
    fn maybe_main() -> Option<Self::PID>;

    fn name() -> &'static str;

    fn process_field_names() -> Vec<Vec<&'static str>>;

    fn process_field_types() -> Vec<Vec<&'static str>>;

    fn process_field_defaults() -> Vec<Vec<&'static str>>;

    fn process_result_types() -> Vec<&'static str>;

    fn process_result_defaults() -> Vec<&'static str>;

    fn channel_local_types() -> Vec<&'static str>;

    fn def() -> Result<Def<Self>, Vec<DefineError>> { ... }
}

Trait specifying types in session context with a method to attempt to create a valid session def struct from those types.

Associated Types

type MID: Id

type CID: Id<Self>

type PID: Id<Self>

type GMSG: Global<Self>

The global message type.

type GPROC: Global<Self>

The global process type.

type GPRES: Global<Self>

The global process result type.

Required Methods

fn maybe_main() -> Option<Self::PID>

fn name() -> &'static str

fn process_field_names() -> Vec<Vec<&'static str>>

fn process_field_types() -> Vec<Vec<&'static str>>

fn process_field_defaults() -> Vec<Vec<&'static str>>

fn process_result_types() -> Vec<&'static str>

fn process_result_defaults() -> Vec<&'static str>

fn channel_local_types() -> Vec<&'static str>

Provided Methods

fn def() -> Result<Def<Self>, Vec<DefineError>>

Return a session def struct if the defined context is valid.

Errors

Process sourcepoints do not correspond one-to-one with channel producers:

#![feature(const_fn)]
#![feature(try_from)]
#[macro_use] extern crate macro_attr;
#[macro_use] extern crate enum_derive;
#[macro_use] extern crate enum_unitary;

extern crate num;
extern crate vec_map;
#[macro_use] extern crate apis;

def_session! {
  context Mycontext {
    PROCESSES where
      let _proc       = self,
      let _message_in = message_in
    [
      process A () {
        kind { apis::process::Kind::isochronous_default() }
        sourcepoints [X]
        endpoints    []
        handle_message { apis::process::ControlFlow::Break }
        update         { apis::process::ControlFlow::Break }
      }
      process B () {
        kind { apis::process::Kind::isochronous_default() }
        sourcepoints []
        endpoints    [X, Y]
        handle_message { apis::process::ControlFlow::Break }
        update         { apis::process::ControlFlow::Break }
      }
    ]
    CHANNELS  [
      channel X <T> (Simplex) {
        producers [A]
        consumers [B]
      }
      channel Y <T> (Sink) {
        producers [A]
        consumers [B]
      }
    ]
    MESSAGES [
      message T {}
    ]
  }
}

fn main() {
  use apis::session::Context;
  assert_eq!(
    Mycontext::def(),
    Err (vec![apis::session::DefineError::ProducerSourcepointMismatch]));
}

Process endpoints do not correspond one-to-one with channel consumers:

#![feature(const_fn)]
#![feature(try_from)]
#[macro_use] extern crate macro_attr;
#[macro_use] extern crate enum_derive;
#[macro_use] extern crate enum_unitary;

extern crate num;
extern crate vec_map;
#[macro_use] extern crate apis;

def_session! {
  context Mycontext {
    PROCESSES where
      let _proc       = self,
      let _message_in = message_in
    [
      process A () {
        kind { apis::process::Kind::isochronous_default() }
        sourcepoints [X,Y]
        endpoints    []
        handle_message { apis::process::ControlFlow::Break }
        update         { apis::process::ControlFlow::Break }
      }
      process B () {
        kind { apis::process::Kind::isochronous_default() }
        sourcepoints []
        endpoints    [X]
        handle_message { apis::process::ControlFlow::Break }
        update         { apis::process::ControlFlow::Break }
      }
    ]
    CHANNELS  [
      channel X <T> (Simplex) {
        producers [A]
        consumers [B]
      }
      channel Y <T> (Sink) {
        producers [A]
        consumers [B]
      }
    ]
    MESSAGES [
      message T {}
    ]
  }
}

fn main() {
  use apis::session::Context;
  assert_eq!(
    Mycontext::def(),
    Err (vec![apis::session::DefineError::ConsumerEndpointMismatch]));
}